[OU-FIX] account: Avoid empty string on name
There can be the possibility of having an empty string in any of the
sources of the field `name`. If that's the case, it will be populated
with such data, leading to not being able to save the generated invoice
through UI until putting '/' in the DB.

With this trick, we avoid this circumstance.

TT32007