[13.0][OU] website_slides: categories migration
The fields `is_preview` and `is_published` are values expected be `True`
as default when the category is created. It can be seen how in the
`create()` and in the `write()` this is explicitly evaluated as long as
in the controller, which is the main way to create categories now. Also,
we must ensure that those categories are active. Al should be as the
old specific model didn't have the field itself.

TT32244