[FIX] calendar: an KeyError occurred while creating recurrent events
Steps:
1. Create an meeting activity that link to recurrent event on
project.task
2. Run migration
3. An KeyError occurred while creating recurrent events

Because In v14, now all occurrences of recurrent events are created as
real records, not virtual ones, so we need to regenerate them for all
the existing ones.
But while creating real records, Odoo will try to create an activity on
the related record. But project.task is not in self.env because the
module calendar doesn't depend on the module project, so models not
installed (is not in self.env)

Solution: we do not create an activity on the real records. An activity that has been link to real record.