[OU-FIX] fleet: Proper link to included services
The field selected for filling the included services was not correct,
as there are 2 one2many to fleet.vehicle.cost, one for the periodically
genereted costs (through the inverse `contract_id`), and another due to
the inherited by delegation one2many through the inverse `parent_id`.

We need to use the second one, which is the correct.