[OU-IMP] account: Empty move partner on regular entries
On pre-v13, account.move>partner_id was a related stored field that
pointed to the first partner found on lines, but now on v13, this field
points only to the partner of the invoices, being empty when you fill a
regular entry.

For consistency, and also for avoiding problems like duplicating a
journal entry, change the parter on the lines, and get a different
partner on header vs lines, we should empty the partner in all the
existing regular entries.

TT38085