[OU-IMP] point_of_sale: stock_location_id
In v13 the deafault location of the picking type is used to take the
stock in the picking operations of the pos configs. In the migration,
this can lead to pickings that were using a common operation type taking
the stock from the wrong locations. We prevent it creating new picking
types for that configs.

TT23199