[OU-FIX] project: complete project visibility when 'followers' option
We forgot to consider two cases in which the user could see the project
in case of 'followers' option was selected.

Thus, we add the channel part. We don't add the task follower part,
because that would make the task follower be able to see all the tasks
of the project, but instead we add the task follower in the task
visibility.