[OU-FIX] account: Switch state to cancel on dangling payments
Coming from older versions of Odoo, you can have a payment with state
"validated" or "reconciled", but no linked move, as it was deleted
through the regular journal entry list.

On that situation, we are going to recreate the moves the same, but
putting them on cancel state for not altering the accounting.

TT39954