[OU-FIX] survey: Change noudpate data on renamed records + no rename
We renamed certain noupdate rules on the migration scripts, but there
was no update on the data inside the records, that change as well. Let's
fix it.

There's also a renaming on a mail template that it's very different, so
let's simply not rename it and remove the old one.