[OU-FIX] account: Don't create counterpart payments for internal transfers
Accounting can't be changed for past internal payments. They will create
extra journal entries that alter the accounting.

On v14, doing an internal transfer payment only puts one journal entry,
and you had to manually create the counterpart internal transfer
payment, but it's not mandatory (as you can do it through other way).
Some code can be added to guess and link both payments if exist, but
the added value is very low, and there are chances of being wrong. So
it's not going to be implemented for now.

On <=v13, creating an internal transfer payment generated 2 journal
entries, so for these payments, an enveloppe payment linked to the
second move can be done, but again the added value is very low, and the
noise on the extra created payments doesn't worth, so not being
implemented either.