[OU-ADD] website_blog: update cover properties field
In v13, blog posts without an image in the header were indicated with
the parameters "backgroung-image" : "none" and "background-color" : "value"
stored in the blog_post.cover_properties field.
When migrating to v14 this style setting is indicated by
"backgroung-image" : "none" and "backgroung_color_class" : "value"
and "background-color" disappears.
For this reason the old publications do not adopt the corresponding
style on the website if they use colour and not an image in their header.
This is solved by rewriting the data contained in the field
blog_post.cover_properties.

TT39016