[OU-FIX] account: don't set term_type
The intention of this script is to render the contents of the
invoice_terms_html field but if we change the company terms type as well
we're going to be forcing a different default from the one expected. For
example, we won't see the note in the sale portal view and instead we'll
see the company terms.

TT44815