[FIX] stock: Incoming moves visibility window
Odoo now uses a global config parameter (or the delay field in stock.rule)
for determining the maximum window for searching for incoming moves when applying
reordering rules.

For preserving the previous behavior, and making sure no delay was previously set
(as in that case, in v14 the showing window is applied using it), we set a high
global window delay.

TT45014