[15.0][OU-ADD] pos_order_mgmt: Merged in point_of_sale TT45178
This module offers the option to see past orders in the fronted, so it is possible to operate with them for re-printing them or performing a return .
This is already accomplished in odoo 15 core, with extra functionallities. To acces them you must select you're desired order in the "Orders" tab of the frontend,
and on the panel there will be the buttons that allows you to perform this actions.
There is one function that is not present in core, which is the option of "duplicate/copy" an order.
By using this function, when it's button is pressed the exact products and quantities of the order are setted in order to be used for another order.
This could justify the creation of a specific module for preserving this feature.