[15.0][OU-ADD] pos_order_return: Merged in point_of_sale TT45179
The OCA module pos_order_return allows the user to partially return a
sale_order in the frontend, and also links this refunds with their
original orders. From Odoo v15 we can do this in core from the frontend
and those links are paired to the ones we were providing.
One thing that adds this module that isn't in core is a boolean indicating
if you want to allow negative quantities on the POS for products.
This feature can be discussed if it is worth it to be conserved
in a custom module or not as it was a little out of context anyway.