[OU][IMP] crm: crm.team use_leads check
Since v15 there are mechanisims to toggle this crm.team flag depending on the
groups and reset some defaults [0]. There's also a mechanisim to automatically set
as lead the crm.lead created from the website form when the assigned team has
the flag use_leads on [1] making that the user wouldn't be able to reach those leads
as the menu isn't available. As we could have inconsistent states from the previous
installation we try to settle the states correctly.

[0] odoo/odoo/blob/15.0/addons/website_crm/models/crm_lead.py#L55
[1] odoo/odoo/tree/15.0/addons/crm/models/res_config_settings.py#L136-L141

TT45672