[OU-ADD] website_crm: Migrate website_crm views
Migrates website_crm views to replace references to 'mail.mail' with 'crm.lead'.
In versions 13 and 14, the creation of opportunities was configured by replacing
the default form view with one that had a specific action to create a lead.
In version 15, this configuration is no longer present, and the default
behavior is to use the base view for sending emails.
To ensure a seamless transition and retain the previous configuration of
creating opportunities, this migration script modifies the relevant views
associated with the website_crm module. It replaces instances of 'mail.mail'
with 'crm.lead', preserving the user's preference for creating opportunities.

TT35811