[OU-ADD] sale_coupon_selection_wizard: Merged into to sale_loyalty_order_suggestion
The wizard module used to allow you to select a promotion and configure it to
be applicable to the sales order. This wizard was opened by a gift icon that
added the suggestion module to the sales order line whose product was part
of the rules of a promotion. The wizard module is no longer needed because
odoo has incorporated a wizard to select promotions. The part of the
functionality to configure the promotion to be applicable that the module
sale_coupon_selection_wizard used to provide has been incorporated into
the module sale_loyalty_order_suggestion adapting it to
the odoo wizard.

TT44352