[OU-FIX] website_sale: Remove views inheritance
On v15, some views were inherited ones, but this it not anymore in v16.
As Odoo only touches the fields that are present in the definition, and
now the `inherit_id` field is missing, the inheritance keeps there,
provoking a crash.

Let's empty them manually by SQL (for avoiding view validation) and
using the key, for any possible website specific view.