[FIX] fill new mail.template template_fs field
the mail.template template_fs field must contain the path of the data
file where it is defined (to allow for it to be reset to its default value).
fill empty values by searching for the mail templates (by xml id) in the xml
files of the module that defines it.