[OU-IMP] account: Rectifiy use_in_tax_closing in old repartition lines
This field was introduced in v14, but it was not impacting in anything
noticeable till this version, where not having this marked in the taxes
lines makes that the tax lines take the analytic dimensions no matter if
the analytic field is marked or not.

As a compromise solution, let's assign this as True for those that have
no value, which are those coming from old versions.