[OU-FIX] hr_expense: Get payment_state directly
Previous approach was too much aggressive, calling a very generic method
that does a lot of stuff apart from recomputing payment_state, launching
a rewrite of the lines with debit=0 and credit=0, which totally wrecks
the accounting.

Now we have extracted the code for computing the payment state inside
the migration script and manually written for each of the expenses
moves.