[OU-ADD] base: web_disable_export_group migration
Take the users from the web_disable_export_group.group_export_data group
to the new base.group_allow_export which covers the same features.

As now Odoo supports disabling data exports, which is the main feature that
the module web_disable_export_group provided we want to move the existing users
in that module's groups to the new core one.

The module's functionality isn't completly merged into core as it allows to
differentiate which type of export users can use.

This might be unnecessary for some module users that would drop the module while
others might want to keep it. To make the transition transparent for both cases,
we put this migration script here.

TT49468