[FIX] stock: stock.move.line creation without reserved quantity
Although this one doesn't prevent the full usage of the module, this
data is not consistent, as all of these move lines are marked as
having reserved quantity, while that doesn't happen when the move is
done, as you can see in [1], and it's transferred to the computed
field [2].

Thus, let's simply put 0 in those fields.

I have discovered it trying to run the
`stock_quant_stock_move_line_desynchronization` action server and
seeing a lot of elements to process.

TT50733

[1] https://github.com/OCA/OpenUpgrade/blob/af2ec6e75e37bf68953148b2024aeefd3a5fa667/addons/stock/models/stock_move_line.py#L460
[2] https://github.com/OCA/OpenUpgrade/blob/af2ec6e75e37bf68953148b2024aeefd3a5fa667/addons/stock/models/stock_move_line.py#L72