[OU-IMP] mail: Compatibility with v15 private channels
In v15 you could set a private group that only members could access, now in v16
this is not possible.
Defining the group_public_id empty will make everyone can access, and leaving the
value that has group_public_id with 'base.group_user' by default is totally wrong.
In order to maintain the compatibility of the existing channels, two transformations
are necessary:
- If a channel is private, but had only one auto-subscription group (group_ids), those
group become the authorized one (group_public_id) of the channel, and it remains as such
as a channel, since at the end, the auto-subscription groups adds as members all the
users of those groups, acting in the same way.
- If the list of members was manual, then the channel becomes a 'Direct Messages'
(channel_type='group'), keeping the channel name and authorized members.

TT50831